# backend_express/src/server.js
import express from 'express';
import cors from 'cors';
import helmet from 'helmet';
import compression from 'compression';
import rateLimit from 'express-rate-limit';
import dotenv from 'dotenv';
import { createServer } from 'http';
import { WebSocketServer } from 'ws';

// Import routes
import telemetryRoutes from './routes/telemetry.js';
import dispatchRoutes from './routes/dispatch.js';
import nodesRoutes from './routes/nodes.js';
import aggregateRoutes from './routes/aggregate.js';
import authRoutes from './routes/auth.js';
import forecastRoutes from './routes/forecast.js';
import optimizationRoutes from './routes/optimization.js';
import marketRoutes from './routes/market.js';
import agentRoutes from './routes/agents.js';
import sshRoutes from './routes/ssh.js';
import vendorAuthRoutes from './routes/vendorAuth.js';
import marketplaceRoutes from './routes/marketplace.js';

// Import services
import { connectDB } from './services/database.js';
import { connectRedis, getRedisClient } from './services/redis.js';
import { initMQTT } from './services/mqtt.js';
import logger from './utils/logger.js';
import { setupWebSocket } from './services/websocket.js';
import forecastScheduler from './jobs/forecastScheduler.js';
import optimizationScheduler from './jobs/optimizationScheduler.js';
import agentScheduler from './jobs/agentScheduler.js';

dotenv.config();

const app = express();
const PORT = process.env.PORT || 3000;
const WS_PORT = process.env.WS_PORT || 3001;

// Middleware
app.use(helmet());
app.use(cors());
app.use(compression());
app.use(express.json());
app.use(express.urlencoded({ extended: true }));

// Rate limiting
const limiter = rateLimit({
  windowMs: 15 * 60 * 1000, // 15 minutes
  max: 100 // limit each IP to 100 requests per windowMs
});
app.use('/api/', limiter);

// Health check
app.get('/health', (req, res) => {
  res.json({
    status: 'healthy',
    timestamp: new Date().toISOString(),
    uptime: process.uptime()
  });
});

// Public routes
app.use('/api/auth', authRoutes);
app.use('/api/vendor-auth', vendorAuthRoutes); // Vendor authentication (JWT-based)

// Public routes - telemetry (no auth required)
app.use('/api/telemetry', telemetryRoutes);

// Public read-only routes (for monitoring/dashboard - no auth required)
app.use('/api/aggregate', aggregateRoutes);
app.use('/api/nodes', nodesRoutes);
app.use('/api/forecast', forecastRoutes);

// Protected write routes (require authentication for modifications)
app.use('/api/dispatch', dispatchRoutes); // Auth handled in route with checkRole
app.use('/api/optimization', optimizationRoutes); // Auth handled in route with checkRole
app.use('/api/market', marketRoutes); // Auth handled in route with checkRole
app.use('/api/agents', agentRoutes); // Multi-agent system routes
app.use('/api/ssh', sshRoutes); // SSH access to edge machines
app.use('/api/marketplace', marketplaceRoutes); // Marketplace routes (bids, transactions, trading)

// Error handling middleware
app.use((err, req, res, next) => {
  // Handle authentication errors from express-jwt
  if (err.name === 'UnauthorizedError' || err.name === 'JsonWebTokenError') {
    logger.warn(`Authentication error: ${err.message}`);
    return res.status(401).json({
      error: 'Unauthorized',
      message: err.message || 'Invalid or missing authentication token',
      timestamp: new Date().toISOString()
    });
  }

  // Handle other errors
  logger.error('Error:', err);
  res.status(err.status || 500).json({
    error: err.message || 'Internal Server Error',
    timestamp: new Date().toISOString()
  });
});

// 404 handler
app.use((req, res) => {
  res.status(404).json({
    error: 'Route not found',
    path: req.path
  });
});

// Initialize services and start server
async function startServer() {
  try {
    // Connect to MongoDB
    await connectDB();
    logger.info('MongoDB connected');

    // Connect to Redis
    await connectRedis();
    logger.info('Redis connected');

    // Initialize MQTT
    await initMQTT();
    logger.info('MQTT client initialized');

    // Check SSH encryption key
    if (!process.env.SSH_ENCRYPTION_KEY || process.env.SSH_ENCRYPTION_KEY === 'default-key-change-in-production-32chars!!') {
      logger.warn('⚠️  SSH_ENCRYPTION_KEY not set or using default! Run: node scripts/generate-ssh-key.js');
      logger.warn('⚠️  SSH credentials will not be properly encrypted in production!');
    } else {
      logger.info('✓ SSH encryption key configured');
    }

    // Start scheduled jobs
    forecastScheduler.start();
    optimizationScheduler.start();
    agentScheduler.start();
    logger.info('Scheduled jobs started (forecast, optimization, agents)');

    // Start HTTP server
    const server = createServer(app);
    server.listen(PORT, () => {
      logger.info(`🚀 Aggregator API running on port ${PORT}`);
    });

    // Start WebSocket server
    const wsServer = createServer();
    const wss = new WebSocketServer({ server: wsServer });
    setupWebSocket(wss);
    
    wsServer.listen(WS_PORT, () => {
      logger.info(`🔌 WebSocket server running on port ${WS_PORT}`);
    });

    // Graceful shutdown
    process.on('SIGTERM', async () => {
      logger.info('SIGTERM received, shutting down gracefully');
      
      // Stop schedulers
      forecastScheduler.stop();
      optimizationScheduler.stop();
      agentScheduler.stop();
      
      // Close SSH connections
      const sshManager = (await import('./services/ssh/sshManager.js')).default;
      await sshManager.closeAllConnections();
      
      server.close(() => {
        logger.info('HTTP server closed');
      });
      wsServer.close(() => {
        logger.info('WebSocket server closed');
      });
      const redis = getRedisClient();
      await redis.quit();
      process.exit(0);
    });

  } catch (error) {
    logger.error('Failed to start server:', error);
    process.exit(1);
  }
}

startServer();