# backend_express/swagger.yaml
openapi: 3.0.3
info:
  title: VPP Aggregation Platform API
  description: |
    API documentation for the Virtual Power Plant (VPP) Aggregation Platform.
    This platform manages edge nodes, telemetry data, dispatch commands, market trading, and multi-agent coordination.
  version: 1.0.0
  contact:
    name: API Support
    email: amansharma12607@gmail.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.vpp-platform.com/api
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login or /api/vendor-auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
        user:
          type: object
          properties:
            email:
              type: string
            name:
              type: string
            role:
              type: string
              enum: [admin, operator, viewer]
...
    VendorRegisterRequest:
      type: object
      required:
        - email
        - password
        - companyName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
    Vendor:
      type: object
      properties:
        email:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
    TelemetryData:
      type: object
      required:
        - dc_id
      properties:
        dc_id:
          type: string
        timestamp:
          type: string
          format: date-time
        power_kw:
          type: number
        voltage_v:
          type: number
        frequency_hz:
          type: number
        battery_soc:
          type: number
        temperature_c:
          type: number
        load_factor:
          type: number
    Node:
      type: object
      required:
        - dc_id
      properties:
        dc_id:
          type: string
        name:
          type: string
        region:
          type: string
        capacity_kw:
          type: number
        battery_capacity_kwh:
          type: number
        location:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_state:
          type: object
        online:
          type: boolean
    DispatchRequest:
      type: object
      required:
        - targets
        - action
      properties:
        targets:
          oneOf:
            - type: string
              enum: [all]
            - type: array
              items:
                type: string
            - type: string
        action:
          type: string
          enum: [charge, discharge, defer_load, hold]
        params:
          type: object
    DispatchResponse:
      type: object
      properties:
        success:
          type: boolean
        dispatched:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        timestamp:
          type: string
          format: date-time
...
    AggregateData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        node_count:
          type: integer
        total_capacity_kw:
          type: number
        total_battery_kwh:
          type: number
        total_power_kw:
          type: number
        avg_soc:
          type: number
        avg_freq:
          type: number
        online_nodes:
          type: integer
        freq_status:
          type: string
        available_reserve_mw:
          type: number
        committed_reserve_mw:
          type: number
        location_flexibility:
          type: object
        nodes:
          type: array
          items:
            type: object
        revenue_today:
          type: number
        co2_saved:
          type: number
...
    DispatchPlan:
      type: object
      properties:
        plan:
          type: object
        expected_revenue:
          type: number
        confidence:
          type: number
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /vendor-auth/register:
    post:
      tags:
        - Vendor Auth
      summary: Register new vendor
      description: Register a new vendor account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorRegisterRequest'
      responses:
        '201':
          description: Vendor registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  vendor:
                    $ref: '#/components/schemas/Vendor'
        '409':
          description: Vendor already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /vendor-auth/login:
    post:
      tags:
        - Vendor Auth
      summary: Vendor login
      description: Authenticate vendor and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  vendor:
                    $ref: '#/components/schemas/Vendor'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /telemetry:
    post:
      tags:
        - Telemetry
      summary: Submit telemetry data
      description: Receive telemetry data from edge nodes
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryData'
      responses:
        '201':
          description: Telemetry data processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - Telemetry
      summary: Get latest telemetry
      description: Retrieve latest telemetry data
      security: []
      parameters:
        - name: dc_id
          in: query
          schema:
            type: string
          description: Filter by data center ID
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of records to return
      responses:
        '200':
          description: Telemetry data retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryData'
...
  /nodes:
    get:
      tags:
        - Nodes
      summary: List all nodes
      description: Get a list of all registered edge nodes
      security: []
      responses:
        '200':
          description: List of nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Node'

    post:
      tags:
        - Nodes
      summary: Register new node
      description: Register a new edge node
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Node'
      responses:
        '201':
          description: Node registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Node'
        '409':
          description: Node already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dispatch:
    post:
      tags:
        - Dispatch
      summary: Dispatch command to edge nodes
      description: Send dispatch commands to one or more edge nodes
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRequest'
      responses:
        '200':
          description: Command dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /aggregate:
    get:
      tags:
        - Aggregation
      summary: Get aggregated virtual plant statistics
      description: Get aggregated statistics for the virtual power plant
      security: []
      parameters:
        - name: region
          in: query
          schema:
            type: string
          description: Filter by region
        - name: nodeIds
          in: query
          schema:
            type: string
          description: Comma-separated list of node IDs
      responses:
        '200':
          description: Aggregated data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  cached:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AggregateData'
...
  /optimization/dispatch:
    post:
      tags:
        - Optimization
      summary: Optimize and execute dispatch
      description: Generate optimized dispatch plan and optionally execute it
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                auto_execute:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Dispatch optimized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/DispatchPlan'
...